Improve repeating payload option
Expand the repeating payload option to more values (char set between
48 and 122) and add 0xdead at the end of each iteration. In a 1500
byte tcp packet the payload aligns to the same 0x10 boundary only
twice. Also, write directly to the buffer to be used in tests vswriting to the fd.

In addition, move the call to fill_with_repeating_pattern to allow
repeating payload to work with legacy -Z / sendfile option.

Signed-off-by: David Ahern <dsahern@gmail.com>